Creating history2mysql
This is working for me but the SQL INSERTs are being one at a time which takes forever.  I want to do them in bulk but I was having trouble doing that.  I'm going to check the script in as-is so I can revert back to this behavior if necessary even though I don't like it.