Support for completed games
The first time I tried to use my tool on a completed game, it had a problem figuring out all the clues for a word.  It turns out that the log lines of a completed game are reversed (oldest first) as compared to the log lines of an active game (newest first).  I changed the script to detect this and reverse the log so it was a very simple change to the script.

I came up with an alternate set of instructions since getting the log is a little different depending on whether the game is active or complete.